fix(security): resolve CodeQL code scanning findings
Address true-positive code scanning alerts:
- Neutralize log injection by stripping CR/LF from interpolated values
  before logging (CLI fatal error and graph ingest failure logs).
- Bound unsigned-to-int limit conversions in unsigned space to avoid
  incorrect narrowing conversions (MCP limits, endpoint vuln limit).
- Handle Close error on a writable file to avoid silent data loss.
- Remove a dead store of recovered validation in graph agent stream.
- Python: drop unused import, type a no-return helper as NoReturn, and
  consolidate a dual module import in the Jira SDK test.

Adds regression tests for limit clamping and log-value sanitization.

+++ internal/bootstrap/mcp.go
@@ -1037,7 +1037,7 @@ func (app *App) mcpInvestigationContext(r *http.Request, args map[string]any) (a
 	if rawLimit, err := mcpUint32Arg(args, "asset_limit"); err != nil {
 		return nil, err
 	} else if rawLimit != 0 {
-		assetLimit = min(int(rawLimit), maxMCPAssetLimit)
+		assetLimit = int(min(rawLimit, uint32(maxMCPAssetLimit)))
 	}
 	assets := []mcpAssetSearchResult{}
 	neighborhoods := []any{}
@@ -2046,7 +2046,7 @@ func mcpBoundedLimit(args map[string]any, key string, defaultLimit int, maxLimit
 	switch {
 	case limit == 0:
 		return defaultLimit, nil
-	case int(limit) > maxLimit:
+	case limit > uint32(maxLimit):
 		return maxLimit, nil
 	default:
 		return int(limit), nil
@@ -2412,7 +2412,7 @@ func mcpNormalizeLimitValue(limit uint32, defaultLimit int, maxLimit int) int {
 	switch {
 	case limit == 0:
 		return defaultLimit
-	case int(limit) > maxLimit:
+	case limit > uint32(maxLimit):
 		return maxLimit
 	default:
 		return int(limit)
